- method opimization for TransactionVO - modification of patient invoice models for Mali - correction of bug (wrong paper size) when printing payment receipts - removal of oc_debet_patientuid index on oc_debets - creation of oc_debet_updatetime index on oc_debets - update medical imaging API - minor bug corrections in DHIS2Exporter - update MTN MoMo payment to accomodate Rwandan platform - correction of date/timestamp error in pharmacy module causing negative stock values in specific cases - update of registers & reporting for Senegal - update of download statistics for compatibility with MySQL 8.0 - additional insurance now also considered in automatic invoicing - added fast payment option to drugs deliverDrugsOutBarcode.jsp - updated paymentCompact.jsp behavior - updated posting of performance data to Global Health Barometer - updated Covid-19 data entry screens - added mandatory field logic to manageRMHOutpatientFile.jsp and manageGitegaOutpatientFileNew.jsp - added recovery of last biometrics in manageRMHOutpatientFile.jsp and manageGitegaOutpatientFileNew.jsp - updated managePsychologyFollowUp.jsp to accept up to 5000 characters of text - minor update getWeightForHeight.jsp - preventing modification of pharmacy kits as long as stock level >0 - updated Wizzeye module - updated financial API - added system administration dahsboard to menu - added mental health screening data entry screens based on CLEFS logic - added functionality for fast updating products list in pharmacy stock - added monitor for MySQL database replication

+++ web/pharmacy/ajax/loadKitProducts.jsp
@@ -3,6 +3,7 @@
 <%
 	StringBuffer sResult=new StringBuffer();
 	String sEditTargetProductStockUid = SH.c(request.getParameter("EditTargetProductStockUid"));
+	ProductStock targetStock = ProductStock.get(sEditTargetProductStockUid);
 	Connection conn = MedwanQuery.getInstance().getOpenclinicConnection();
 	PreparedStatement ps = conn.prepareStatement("select sum(OC_PRODUCTKIT_SOURCEPRODUCTSTOCKQUANTITY) OC_PRODUCTKIT_SOURCEPRODUCTSTOCKQUANTITY,OC_PRODUCTKIT_SOURCEPRODUCTSTOCKUID from oc_productkits where OC_PRODUCTKIT_TARGETPRODUCTSTOCKUID=? group by OC_PRODUCTKIT_SOURCEPRODUCTSTOCKUID order by oc_productkit_sourceproductstockuid");
 	ps.setString(1, sEditTargetProductStockUid);
@@ -10,7 +11,12 @@
 	while(rs.next()){
 		ProductStock stock = ProductStock.get(rs.getString("OC_PRODUCTKIT_SOURCEPRODUCTSTOCKUID"));
 		int quantity = rs.getInt("OC_PRODUCTKIT_SOURCEPRODUCTSTOCKQUANTITY");
-		sResult.append("<tr><td width='1%'><img onclick='deleteKitProduct(\""+stock.getUid()+"\")' src='"+sCONTEXTPATH+"/_img/icons/icon_delete.png'/></td><td width='1%' nowrap>"+quantity+"</td><td>x <b>"+stock.getProduct().getName()+"</b> ["+stock.getServiceStock().getName()+"]</td></tr>");
+		if(activeUser.getAccessRightNoSA("pharmacy.kits.edit") || targetStock.getLevel()==0){
+			sResult.append("<tr><td width='1%'><img onclick='deleteKitProduct(\""+stock.getUid()+"\")' src='"+sCONTEXTPATH+"/_img/icons/icon_delete.png'/></td><td width='1%' nowrap>"+quantity+"</td><td>x <b>"+stock.getProduct().getName()+"</b> ["+stock.getServiceStock().getName()+"]</td></tr>");
+		}
+		else{
+			sResult.append("<tr><td width='1%'></td><td width='1%' nowrap>"+quantity+"</td><td>x <b>"+stock.getProduct().getName()+"</b> ["+stock.getServiceStock().getName()+"]</td></tr>");
+		}
 	}
 	rs.close();
 	ps.close();
